+++ lib/db/src/schema/index.ts
@@ -55,3 +55,4 @@ export * from "./eenheidsprijzen";
 export * from "./fie";
 export * from "./avg";
 export * from "./pim";
+export * from "./kb";